requirements.txt

# Data processing
pandas>=2.1.0
pyarrow>=14.0.0
pyspark>=3.5.0

# ML
lightgbm>=4.1.0
scikit-learn>=1.3.0
shap>=0.43.0

# Graph
networkx>=3.2
python-louvain>=0.16

# Agent & LLM
langgraph>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0

# MCP server
fastmcp>=0.1.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0

# Kafka
kafka-python>=2.0.2

# Dashboard
streamlit>=1.29.0
pyvis>=0.3.2

# Experiment tracking
mlflow>=2.9.0

# Data generation
faker>=20.0.0

# Templating (for report generator)
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0